Fine tuning/code/requirements.txt
# requirements.txt
# Install in Colab with: !pip install -r requirements.txt

torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
tokenizers>=0.14.0
datasets>=2.14.0
lmdb>=1.4.1
Pillow>=9.5.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
albumentations>=1.3.1
jiwer>=3.0.3
pyyaml>=6.0
tqdm>=4.65.0
